Extract getModuleSpecifier into its own module
