+++ lib/get-module-config.js
@@ -1,4 +1,4 @@
-"use strict";
+'use strict';
 
 module.exports = function(config) {
   let moduleConfig = config.moduleConfiguration;
